Do not trip on new warn before read feature
Upstream privatebin introduced this new feature  https://github.com/PrivateBin/PrivateBin/pull/1237

If the fragment contains a prefix character `-` , then the ui is supposed to warn before actyally getting paste, as the paste may be burn on read.

This character can trip our reading of the bs58key. So, strip the prefix if found.